kill YamlGlobalConfig, merge into BaseGlobalConfig
