+++ cumulusci/core/config/BaseGlobalConfig.py
@@ -1,16 +1,25 @@
 from __future__ import unicode_literals
 import os
 
-from cumulusci.core.config.BaseProjectConfig import BaseProjectConfig
+import yaml
+from cumulusci.core.config.YamlProjectConfig import YamlProjectConfig
+from cumulusci.core.config.MergedConfig import MergedConfig
 from cumulusci.core.config import BaseTaskFlowConfig
+
+__location__ = os.path.dirname(os.path.realpath(__file__))
 
 
 class BaseGlobalConfig(BaseTaskFlowConfig):
     """ Base class for the global config which contains all configuration not specific to projects """
 
-    project_config_class = BaseProjectConfig
-
+    config_filename = "cumulusci.yml"
+    project_config_class = YamlProjectConfig
     config_local_dir = ".cumulusci"
+
+    def __init__(self, config=None):
+        self.config_global_local = {}
+        self.config_global = {}
+        super(BaseGlobalConfig, self).__init__(config)
 
     def list_projects(self):
         """ Returns a list of project names """
@@ -23,3 +32,37 @@ class BaseGlobalConfig(BaseTaskFlowConfig):
     def create_project(self, project_name, config):
         """ Creates a new project configuration and returns it """
         raise NotImplementedError("Subclasses must provide an implementation")
+
+    @property
+    def config_global_local_path(self):
+        directory = os.path.join(os.path.expanduser("~"), self.config_local_dir)
+        if not os.path.exists(directory):
+            os.makedirs(directory)
+
+        config_path = os.path.join(directory, self.config_filename)
+        if not os.path.isfile(config_path):
+            return None
+
+        return config_path
+
+    @property
+    def config_global_path(self):
+        return os.path.abspath(
+            os.path.join(__location__, "..", "..", self.config_filename)
+        )
+
+    def _load_config(self):
+        """ Loads the local configuration """
+        # load the global config
+        with open(self.config_global_path, "r") as f_config:
+            config = yaml.safe_load(f_config)
+        self.config_global = config
+
+        # Load the local config
+        if self.config_global_local_path:
+            config = yaml.safe_load(open(self.config_global_local_path, "r"))
+            self.config_global_local = config
+
+        self.config = MergedConfig(
+            global_local=self.config_global_local, global_config=self.config_global
+        ).config
