+++ cumulusci/core/config/BaseProjectConfig.py
@@ -335,7 +335,7 @@ class BaseProjectConfig(BaseTaskFlowConfig):
         """ location of the user local directory for the project
         e.g., ~/.cumulusci/NPSP-Extension-Test/ """
 
-        # depending on where we are in bootstrapping the YamlGlobalConfig
+        # depending on where we are in bootstrapping the BaseGlobalConfig
         # the canonical projectname could be located in one of two places
         if self.project__name:
             name = self.project__name
